print the real line failing in the notebook's flake8 linting
The syntax used to add checks/annotations doesn't show up in the workflow log file.
But in certain circumstances they (maybe a bug in this code) they don't always
show up in the PR either. This change should make it easier to at least infer
what the source of the error is, especially when dealing with large notebooks that
have a syntax error.